// src/models/models.ts
// types.ts
export type Product = {
    id: string;
    name: string;
    type: string;
    description: string;
    memory: string;
    stock: number;
    price: number;
    image_urls: string[];
    brand: string;
    manufacturer: string;
    model: string;
    brand_warranty_months: number;
    EAN: string;
  };
  